requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
gunicorn==21.2.0
httpx==0.27.2
python-dotenv==1.0.0
pydantic==2.10.6
# langchain
langchain-core==0.3.41
langchain-openai==0.3.7
langchain-community==0.3.19
langchain-google-community==2.0.7
openai==1.68.2
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
python-multipart==0.0.20
jinja2==3.1.2
aiofiles==23.2.1
# tavily-python==0.2.6
altair>=5.2.0
# vega>=3.6.0
# vega-lite>=5.7.0
docker==7.0.0
vega_datasets==0.9.0
plotly==5.18.0
yfinance==0.2.54
langgraph==0.3.5
openai-agents==0.0.6
wikipedia==1.4.0
flask-cors==5.0.1
requests==2.31.0
tiktoken==0.8.0
starlette==0.45.3
websockets==14.2
tabulate==0.9.0